Adapted code for tests to run successfully.

+++ index.js
@@ -5,7 +5,10 @@ import express from 'express'
 const app = express()
 const PORT = 8000
 
-sequelize.sync({ force: true }).then(() => console.log('db is ready'))
+
+if (process.env.NODE_ENV !== 'test') {
+    sequelize.sync({ force: true }).then(() => console.log('db is ready'))
+}
 
 app.use(express.json())
 app.use('/api/users', usersRouter)
